* hotfix/7.8.6:
  Control better features with MARS_SERVER switch
  CMake: move ecbuild_declare_project forward
  MARS-593: fix fdb-wipe
  Update sublime text project

+++ src/fdb5/toc/WipeVisitor.cc
@@ -184,7 +184,10 @@ void WipeVisitor::wipe(std::ostream&) const {
         }
     }
 
-    directory.rmdir();
+    if(directory.exists()) {
+        directory.rmdir();
+    }
+
 }
